+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 
 - Internal refactoring.
 
+### `@liveblocks/node`
+
+- Fix: improve stack traces of REST API errors to include the original error
+  location.
+
 ## v2.22.0
 
 ### `@liveblocks/node`